[refactor] 폴더 구조 변경, 파일 위치 변경
enums라는 폴더 만들어 그곳에 enum class 저장

#124
