+++ backend/src/main/java/com/ddbb/dingdong/application/usecase/reservation/GetReservationsUseCase.java
@@ -5,8 +5,8 @@ import com.ddbb.dingdong.application.common.UseCase;
 import com.ddbb.dingdong.domain.reservation.entity.vo.ReservationStatus;
 import com.ddbb.dingdong.domain.reservation.repository.ReservationQueryRepository;
 import com.ddbb.dingdong.domain.reservation.repository.projection.UserReservationProjection;
-import com.ddbb.dingdong.presentation.endpoint.reservation.exchanges.ReservationCategory;
-import com.ddbb.dingdong.presentation.endpoint.reservation.exchanges.SortType;
+import com.ddbb.dingdong.presentation.endpoint.reservation.exchanges.enums.ReservationCategory;
+import com.ddbb.dingdong.presentation.endpoint.reservation.exchanges.enums.SortType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
