--- backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/reservation/exchanges/SortType.java
@@ -1,6 +1,0 @@
-package com.ddbb.dingdong.presentation.endpoint.reservation.exchanges;
-
-public enum SortType {
-    LATEST,
-    OLDEST,
-}
