+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/reservation/exchanges/enums/SortType.java
@@ -1,0 +1,6 @@
+package com.ddbb.dingdong.presentation.endpoint.reservation.exchanges.enums;
+
+public enum SortType {
+    LATEST,
+    OLDEST,
+}
